requirements.txt
Django==1.7.7
Pillow==2.6.1
psycopg2==2.6
djangorestframework==3.1
Markdown==2.6.1
django-oauth-toolkit==0.8.0
django-filter==0.9.2
django-guardian==1.2.5
requests==2.6.0
